// functions/src/utils/enums/actionType.ts
export enum ActionType {
    Video = "Video",
    Quiz = "Quiz",
    Claim = "Claim",
    SocialTwitter = "Social_Twitter",
    SocialDiscord = "Social_Discord",
    Notify = "Notify",
    GenerateInviteCode = "Generate_Invite_Code",
    CheckProfile = "CheckProfile",
    VerifyInviteCode = "Verify_Invite_Code",
    DailyReward = "Daily_Reward",
    BuildProfile = "BuildProfile"
}